# go/cmd/api/docs/swagger.yaml
basePath: /
definitions:
  auth.TokenResponse:
    properties:
      expires_at:
        type: integer
      token:
        type: string
    type: object
  protocol.TaskStatusUpdate:
    properties:
      id:
        type: integer
      mc:
        default: 1
        maximum: 255
        minimum: 1
        type: integer
      mode:
        enum:
        - observation
        - stress
        type: string
      progress:
        default: 0
        type: integer
      sem:
        enum:
        - shared
        - api
        type: string
      status:
        enum:
        - check_lock
        - progress
        - completed
        - lock_acquired
        - lock_failed
        - retries_failed
        - retry
        type: string
      worker:
        default: 0
        type: integer
    type: object
  semaphore.AcquireRequest:
    properties:
      lock_wait_timeout:
        example: 1
        minimum: 0
        type: integer
      max_concurrent:
        example: 5
        maximum: 255
        minimum: 1
        type: integer
      permit_ttl:
        example: 2
        maximum: 60
        minimum: 1
        type: integer
    type: object
  semaphore.ReleaseRequest:
    properties:
      uid:
        example: "5:3"
        type: string
    type: object
info:
  contact: {}
  description: Task status receiver and semaphore API
  title: FX API
  version: "1.0"
paths:
  /auth/token:
    post:
      consumes:
      - application/json
      description: Returns a signed JWT token valid for 24 hours
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: invalid request body
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Generate JWT token
      tags:
      - auth
  /semaphore/acquire:
    post:
      consumes:
      - application/json
      description: Acquires a slot in the distributed semaphore. Returns a unique
        slot UID on success.
      parameters:
      - description: Semaphore acquire request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/semaphore.AcquireRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid JSON body
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "409":
          description: no available slots
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Acquire semaphore slot
  /semaphore/health:
    get:
      description: Returns 200 OK
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Semaphore health check
  /semaphore/release:
    post:
      consumes:
      - application/json
      description: Releases a slot identified by its UID. Idempotent — releasing an
        already released slot is a no-op.
      parameters:
      - description: Slot UID to release
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/semaphore.ReleaseRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: Slot released (or was already free)
        "400":
          description: invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: redis connection error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Release semaphore slot
  /task/status:
    post:
      consumes:
      - application/json
      parameters:
      - description: Task status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/protocol.TaskStatusUpdate'
      responses:
        "202":
          description: Accepted
        "400":
          description: invalid JSON body
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: failed to publish to NATS
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: SendStatus receives task status update and publishes to NATS.
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"