# go/internal/api/task/handlers.go
// Package task provides HTTP handlers for task status forwarding to NATS.
package task

import (
	"encoding/json"
	"fx/go/internal/protocol"
	"net/http"
)

// Publisher defines the interface for publishing messages to NATS.
type Publisher interface {
	Publish(subj string, data []byte) error
}

// Handler forwards task status updates to NATS.
type Handler struct {
	publisher Publisher // nats conn for now
	subject   string    // nats subject for now
}

// NewHandler creates a task handler with the given publisher and subject.
func NewHandler(publisher Publisher, subject string) *Handler {
	return &Handler{
		publisher: publisher,
		subject:   subject,
	}
}

// SendStatus forwards task status from PHP to NATS.
//
// @Summary  SendStatus receives task status update and publishes to NATS.
// @Security ApiKeyAuth
// @Accept   json
// @Param    status body protocol.TaskStatusUpdate true "Task status"
// @Success  202
// @Failure  400 {string} string "invalid JSON body"
// @Failure  401 {string} string "Unauthorized"
// @Failure  500 {string} string "failed to publish to NATS"
// @Router   /task/status [post]
func (h *Handler) SendStatus(w http.ResponseWriter, r *http.Request) {
	var req protocol.TaskStatusUpdate

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "invalid JSON body: "+err.Error(), http.StatusBadRequest)
		return
	}

	// Serialize back to JSON
	// Marshal just the status first
	statusBytes, err := json.Marshal(req)
	if err != nil {
		http.Error(w, "failed to marshal status: "+err.Error(), http.StatusInternalServerError)
		return
	}
	payload, err := json.Marshal(protocol.NatsEnvelope{
		Type: protocol.MsgTypeStatusUpdate,
		Data: statusBytes,
	})
	if err != nil {
		http.Error(w, "failed to marshal task status: "+err.Error(), http.StatusInternalServerError)
		return
	}

	// Publish status to NATS
	if err := h.publisher.Publish(h.subject, payload); err != nil {
		http.Error(w, "failed to publish to NATS: "+err.Error(), http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusAccepted)
}